CASMCMS-7387: Take artifacts from algol60
As part of the migration to Github, CASM supports its artifacts from
its own artifactory instance, algol60. This commit references that
instance.

Create a base layer in the Dockerfile. This layer can be built by
specifying it as a target in the 'docker build --target base' command.
Doing this allows us to remove the now redundant Dockerfile.devel file.

The cms_meta_tools were never used, so the script to install them was
deleted.