feature/CASMCMS-8505: Update base builder image to support arm64 builds
The alpine OS environment does not adequately support cross compile building of arm based ipxe binaries, so we're changing the base environment over to one that does. As a requirement response, downstream building applications (cms-ipxe) will need to change their docker build environment in order to accomodate the different ecosystem.

Patchification to this repository has begun in an effort to move global source modifications into the build image in exactly one place in attempt to make the environment more thread safe, so that multiple binaries can be built simultaneously. It is not immediately clear how much further patchification is required for us to remove all forked source code from our repository, but that is the supported direction that we would like to pursue. To this end, this mod includes some commented code that will require further conversation with our build team to allow checkout of external git sources as a part of overall build from outside sources. It is not currently possible to enable this because secrets and credentials need to be strung through in order for the checkout/clone to work.

This dockerfile includes additional RUN stanzas that exercise a closer to nominal binary build model; this serves two purposes: smoke tests for compile, and precompile of source as a runtime optimization.

This changeset should target a major milestone version to better represent the backwards incompatibilities it introduces.